+++ http-core/src/main/scala/com/twitter/finatra/http/request/RequestBuilder.scala
@@ -1,4 +1,4 @@
-package com.twitter.finatra.httpclient
+package com.twitter.finatra.http.request
 
 import com.twitter.finagle.http.{Fields, Message, Method, Request, RequestProxy}
 import com.twitter.io.StreamIO
@@ -7,65 +7,51 @@ import java.nio.charset.StandardCharsets.UTF_8
 /**
  * Provides a class for building <code>finagle.http.Request</code> objects
  */
-@deprecated("Use com.twitter.finatra.http.request.RequestBuilder directly.", "02-16-2020")
 object RequestBuilder {
-
-  @deprecated("Use com.twitter.finatra.http.request.RequestBuilder.get directly.", "02-16-2020")
   def get(url: String): RequestBuilder = {
     create(Method.Get, url)
   }
 
-  @deprecated("Use com.twitter.finatra.http.request.RequestBuilder.post directly.", "02-16-2020")
   def post(url: String): RequestBuilder = {
     create(Method.Post, url)
   }
 
-  @deprecated("Use com.twitter.finatra.http.request.RequestBuilder.put directly.", "02-16-2020")
   def put(url: String): RequestBuilder = {
     create(Method.Put, url)
   }
 
-  @deprecated("Use com.twitter.finatra.http.request.RequestBuilder.patch directly.", "02-16-2020")
   def patch(url: String): RequestBuilder = {
     create(Method.Patch, url)
   }
 
-  @deprecated("Use com.twitter.finatra.http.request.RequestBuilder.delete directly.", "02-16-2020")
   def delete(url: String): RequestBuilder = {
     create(Method.Delete, url)
   }
 
-  @deprecated("Use com.twitter.finatra.http.request.RequestBuilder.head directly.", "02-16-2020")
   def head(url: String): RequestBuilder = {
     create(Method.Head, url)
   }
 
-  @deprecated("Use com.twitter.finatra.http.request.RequestBuilder.trace directly.", "02-16-2020")
   def trace(url: String): RequestBuilder = {
     create(Method.Trace, url)
   }
 
-  @deprecated("Use com.twitter.finatra.http.request.RequestBuilder.connect directly.", "02-16-2020")
   def connect(url: String): RequestBuilder = {
     create(Method.Connect, url)
   }
 
-  @deprecated("Use com.twitter.finatra.http.request.RequestBuilder.options directly.", "02-16-2020")
   def options(url: String): RequestBuilder = {
     create(Method.Options, url)
   }
 
-  @deprecated("Use com.twitter.finatra.http.request.RequestBuilder.create directly.", "02-16-2020")
   def create(method: Method, url: String): RequestBuilder = {
     new RequestBuilder(Request(method, url))
   }
-
 }
 
 /**
  * RequestBuilder is a finagle.http.Request with a builder API for common mutations
  */
-@deprecated("Use com.twitter.finatra.http.request.RequestBuilder directly.", "02-16-2020")
 class RequestBuilder(override val request: Request) extends RequestProxy {
 
   def headers(headers: Map[String, String]): RequestBuilder = {